refactor tests and file names.
