+++ qtgql/gqltransport/client.py
@@ -12,7 +12,7 @@ from PySide6 import QtNetwork as qtn
 from PySide6 import QtWebSockets as qtws
 
 from qtgql import slot
-from qtgql.gqlcore.gqlcore import EncodeAble, GqlEncoder, QueryPayload, T
+from qtgql.gqltransport.core import EncodeAble, GqlEncoder, QueryPayload, T
 from qtgql.typingref import UNSET
 
 logger = logging.getLogger(__name__)
