+++ tests/test_compiler/test_py/test_introspection_generator.py
@@ -1,0 +1,8 @@
+import pytest
+
+from tests.mini_gql_server import schema
+
+
+@pytest.fixture
+def introspection_query() -> dict:
+    return schema
