+++ tests/test_sample_ui/main.py
@@ -7,7 +7,7 @@ from PySide6 import QtCore, QtGui, QtQml, QtQuick
 from PySide6.QtCore import QObject, Signal
 from PySide6.QtQml import QQmlApplicationEngine
 from qtgql import slot
-from qtgql.gqlcore.client import GqlClientMessage, GqlWsTransportClient, HandlerProto
+from qtgql.gqltransport.client import GqlClientMessage, GqlWsTransportClient, HandlerProto
 from qtgql.itemsystem import GenericModel
 
 from tests.test_sample_ui.models import Apple
